Merged the 4 files into one file.
Changed the proto declaration from xp11in and xp11out to just "xplane" (ie  now filter on xplane.data instead of xp11in.data). There is no longer a differentiation between in and out. To filter on inbound and outbound packet use udp.srcport and udp.dstport or just udp.port.

Removed all the other proto declarations as they were unnecessary and clogging up the proto's list.

Added a Preference for the BECN port (default 49707)
TODO: Add preferences for other ports to help with heuristic filtering.

Changed version to 1.1.0

Added some rudimentary checks to see if we should dissect this packet (returns false better if a header isn't for us).

NFAL/NREC now checked and use a string as I've discovered the whole name of the Navaid is to be passed as a string, not an int). Just leaves ISE6 as not checked

Changed all but the DATA lookups to use wireshark's value_string mechanism. DATA lookup is 2D so still uses a table and :append_text

Now only adds one subtree to the main tree instead of 2 with the packet size on the subtree desc and the header move to the packet content section.